requirements.txt
fastapi
uvicorn
gunicorn
motor
pymongo
pydantic-settings
python-dotenv
python-jose
twilio
email-validator
cloudinary
passlib[bcrypt]
python-multipart
python-socketio[asyncio]
websockets
httpx
redis>=5.0.0
stripe>=8.0.0
python-socketio[asgi]